Revert "moved the stateless functions to session utils module"
This reverts commit 468a6579c278927643125c2313411899dddb320f.
